% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgld-ls-robust.R
\name{assign_diff_lengths}
\alias{assign_diff_lengths}
\title{Assignment problem with two vectors of different size}
\usage{
assign_diff_lengths(x1, x2)
}
\arguments{
\item{x1}{vector/matrix 1}

\item{x2}{vector/matrix 2}
}
\value{
Indices of \code{x2} which minimize the sum of distances to \code{x1}.
}
\description{
Assignment problem with two vectors of different size
}
\examples{
x1 <- c(-10, 3, 4, 5)
x2 <- c(3, 4)
assign_diff_lengths(x1, x2)

}
