% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqclass.R
\name{dqclass_train}
\alias{dqclass_train}
\alias{predict_dqclass}
\alias{dq_train_test}
\title{Integrated rank-weighted maximum depth classifier}
\usage{
dqclass_train(
  X,
  y,
  distr = c("normal", "kde", "fgld"),
  n_dir = 1000,
  S,
  predicted = FALSE,
  weighted = FALSE,
  sphered = TRUE
)

predict_dqclass(out_train, X)

dq_train_test(
  train,
  test,
  n_dir = 500,
  S,
  distr = "fgld",
  weighted = FALSE,
  sphered = FALSE
)
}
\arguments{
\item{X}{A data matrix with observations in rows and variables in columns.}

\item{y}{A vector of class labels for each sample of the training set.}

\item{distr}{The distribution that is fit to each univariate directional projection.}

\item{n_dir}{Number of directions to consider. They are equally spaced in two dimensions and uniformly sampled for higher dimensions.}

\item{S}{A matrix containing the directions in the columns.}

\item{predicted}{A logical indicating whether the predicted values from the training sample should be returned.}

\item{weighted}{A logical indicating whether weights for the directions should be computed.}

\item{sphered}{A logical indicating whether the depth should be computed on the sphered data.}

\item{out_train}{Ouput list from \code{dqclass_train}.}

\item{train, test}{Matrices for training and testing data sets.
The last column must contain the response class coded via integer numbers.}
}
\value{
A list containing:
\itemize{
\item \code{theta}: a list of dimension \code{n_dir} containing the estimated parameters for each univariate directional projection.
\item \code{sphere}: a list of group specific sphering matrices.
}
}
\description{
Integrated rank-weighted maximum depth classifier
}
\examples{
data("iris")
out_train <- dqclass_train(iris[, -5], iris[, 5], predicted = TRUE)
table(iris[, 5], out_train$class)

}
