% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqdepth.R
\name{dqdepth}
\alias{dqdepth}
\title{Directional quantile depth}
\usage{
dqdepth(z, X, S, n_dir = 1000, distr = "normal", weighted = FALSE)
}
\arguments{
\item{z}{Point(s) whose depth is to be computed. If omitted the depths are computed for the points in \code{X}.}

\item{X}{Data matrix with respect to which the depth is to be computed, with observations in rows and variables in columns.}

\item{S}{A matrix containing the directions in the columns.}

\item{n_dir}{Number of directions to consider. They are equally spaced in two dimensions and uniformly sampled for higher dimensions.}

\item{distr}{The distribution that is fit to each univariate directional projection.}

\item{weighted}{A logical indicating whether weights for the directions should be computed.}
}
\value{
A list containing:
\itemize{
\item \code{theta}: a list of dimension \code{n_dir} containing the estimated parameters for each univariate directional projection.
\item \code{depths}: a vector with the computed depths for the input data \code{X}.
}
}
\description{
Computes the depth based on directional quantiles for a data matrix \eqn{X}.
}
\examples{
library(ggplot2)
library(mvtnorm)
set.seed(1)
X <- rmvnorm(n = 200, mean = c(10, 2), sigma = matrix(c(1, 0.9, 0.9, 2), 2, 2))
depths <- dqdepth(X = X, weighted = FALSE)

data.frame("x1" = X[,1], "x2" = X[,2], "d" = depths) |>
  ggplot(aes(x1, x2, col = d))+
  geom_point()+
  scale_color_viridis_c()

}
