% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgld-ls-robust.R
\name{fit_fgld_ls_robust}
\alias{fit_fgld_ls_robust}
\title{Robust least squares estimation for the \emph{fgld}}
\usage{
fit_fgld_ls_robust(y, alpha)
}
\arguments{
\item{y}{Data vector.}

\item{alpha}{Proportion of contamination.}
}
\value{
List containing:
\itemize{
\item \code{theta}: estimated vector of parameters.
\item \code{idx}: indices of the observations that have been identified as outliers.
\item \code{obj}: vector of the objective evaluations at each iteration.
}
}
\description{
Robust least squares estimation for the \emph{fgld}
}
\examples{
set.seed(1)
x <- c(rlogis(18), -10, 10)
fit_fgld_ls_robust(x, alpha = 0.1)

}
