% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgldh-ls-estimation.R
\name{get_X_fgldh}
\alias{get_X_fgldh}
\alias{fit_fgldh_ls}
\title{Least squares estimation for the \emph{fgldh}}
\usage{
get_X_fgldh(n)

fit_fgldh_ls(y)
}
\arguments{
\item{n}{Sample size.}

\item{y}{Data from which to estimate the parameters.}
}
\value{
\code{fit_fgldh_ls()} outputs a vector of length 6 containg the estimates of the parameters.

\code{get_X_fgldh()} outputs a \eqn{n \times 6} matrix where the i-th row contains the
coefficients for expected value of the i-th order statistic of the fgld.
}
\description{
\code{fit_fgldh_ls()} carries out the least squares estimation of the
\emph{fgld} (flattened generalized logistic distribution) with a linear
reparametrization.

\code{get_X_fgldh()} returns the design matrix needed for the least squares estimation. The design matrix contains the coefficients of the
expected order statistics.
}
\examples{
x <- rnorm(10)
# get_X_fgldh(10)
fit_fgldh_ls(x)

}
