% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-ls-estimation.R
\name{get_X_logis}
\alias{get_X_logis}
\alias{fit_logis_ls.quadprog}
\alias{fit_logis_ls}
\alias{fit_logis_ls_obj}
\title{Least squares estimation for the logistic distribution}
\usage{
get_X_logis(n)

fit_logis_ls.quadprog(y)

fit_logis_ls(y)

fit_logis_ls_obj(y)
}
\arguments{
\item{n}{Sample size.}

\item{y}{Data from which to estimate the parameters.}
}
\value{
\code{fit_logis_ls()} outputs a vector of length 2 containg the estimates of the parameters.

\code{get_X_logis()} outputs a \eqn{n \times 2} matrix where the i-th row contains the
coefficients for expected value of the i-th order statistic of the logistic distribution.
}
\description{
Least squares estimation for the logistic distribution
}
\examples{
set.seed(123)
x <- rlogis(100, 3, 2)
fit_logis_ls(x)

}
