% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq-utils.R
\name{get_equi_S}
\alias{get_equi_S}
\title{Equally spaced directions in two dimensions}
\usage{
get_equi_S(B)
}
\arguments{
\item{B}{Number of directions.}
}
\value{
A matrix of dimension \eqn{2 \times B} where each column is a direction in the 2D plane.
}
\description{
Function used for two dimensional samples that creates equally spaced directions by dividing the unit circle in the required number of angles.
}
\examples{
get_equi_S(5)
}
