% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq-utils.R
\name{get_S}
\alias{get_S}
\title{Random directions}
\usage{
get_S(B, dim)
}
\arguments{
\item{B}{Number of random directions to generate.}

\item{dim}{Dimension for each direction.}
}
\value{
A matrix of dimension \eqn{dim \times B} where each column is a direction.
}
\description{
Random directions
}
\examples{
get_S(10, 3)

}
