% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgldh-ls-estimation.R
\name{qfgldh}
\alias{qfgldh}
\alias{qfgldh_deriv}
\alias{pfgldh}
\alias{dfgldh}
\title{The flattened generalized logistic distribution with heavy tails (\emph{fgldh})}
\usage{
qfgldh(u, theta)

qfgldh_deriv(u, theta)

pfgldh(q, theta)

dfgldh(x, theta, log = FALSE)
}
\arguments{
\item{u}{Vector of probabilities.}

\item{theta}{Vector of parameters (of length 4) for the \emph{fgldh}.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical indicating whether the density should be returned on the log scale.}
}
\value{
\code{qfgld} gives the quantile function, \code{pfgld} gives the distribution functions.
The length of the output is the same as that of the input \code{x/u}.
}
\description{
Quantile and cumulative distribution functions for the \emph{fgldh}.
}
\examples{
theta <- c(1, 2, 0, 0, 1, 2)
u <- seq(0.1, 0.9, by = 0.1)
qfgldh(u, theta)

}
