% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq-utils.R
\name{sphere_zca}
\alias{sphere_zca}
\title{Compute a sphering matrix}
\usage{
sphere_zca(X, centered = FALSE)
}
\arguments{
\item{X}{A matrix of dimensions \eqn{n \times p}, where \eqn{n} is the number
of points and \eqn{p} the number of dimensions.}

\item{centered}{A logical indicating whether the data matrix in argument \code{X}
has already been centered.}
}
\value{
A sphering matrix of dimensions \eqn{p \times p} that premultiplied
by data matrix \code{X} gives a sphered data matrix.
}
\description{
Computes a sphering (or whitening) matrix for a data matrix \eqn{X}, that is a matrix \eqn{W}, such that \eqn{Z = X
W} has the identity matrix as sample covariance matrix. Uses the so-called ZCA (zero-phase component analysis) or
Mahalanobis whitening procedure. The whitening matrix is the inverse matrix
square root of the sample covariance matrix \eqn{\hat{\Sigma}}.
\eqn{W^{ZCA} = \hat{\Sigma}^{-\frac{1}{2}}}
The inverse square root is computed via the \code{svd} decomposition of the
centered data matrix.
}
\examples{
library(mvtnorm)
Sigma <- matrix(c(1, 0.9, 0.9, 2), 2, 2)
X <- mvtnorm::rmvnorm(n = 100, sigma = Sigma)
var(X)
W <- sphere_zca(X)
Z <- X \%*\% W
var(Z)

}
