% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqclass.R
\name{dqclass_train}
\alias{dqclass_train}
\title{Directional quantile classifier}
\usage{
dqclass_train(
  X,
  y,
  distr = c("normal", "kde", "fgld"),
  n_dir = 1000,
  S,
  predicted = FALSE,
  weighted = TRUE,
  sphered = TRUE
)
}
\arguments{
\item{X}{A data matrix with observations in rows and variables in columns.}

\item{y}{A vector of class labels for each sample of the training set.}

\item{distr}{The distribution that is fit to each univariate directional projection.}

\item{n_dir}{Number of directions to consider. They are equally spaced in two dimensions and uniformly sampled for higher dimensions.}

\item{S}{A matrix containing the directions in the columns.}

\item{predicted}{A logical indicating whether the predicted values from the training sample should be returned.}

\item{weighted}{A logical indicating whether weights for the directions should be computed.}

\item{sphered}{A logical indicating whether groups should be sphered.}
}
\value{
A list containing:
\itemize{
\item \code{theta}: a list of dimension \code{n_dir} containing the estimated parameters for each univariate directional projection.
\item \code{sphere}: a list of group specific sphering matrices.
}
}
\description{
Directional quantile classifier
}
\examples{
data("iris")
out_train <- dqclass_train(iris[, -5], iris[, 5], predicted = TRUE)
table(iris[, 5], out_train$class)

}
