% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqdepth.R
\name{dqcontour}
\alias{dqcontour}
\title{Directional quantile contours}
\usage{
dqcontour(u, X, S, n_dir = 1000, distr = "normal", weighted = FALSE)
}
\arguments{
\item{weighted}{}
}
\description{
Directional quantile contours
}
\examples{
library(mvtnorm)
set.seed(2)
X <- rmvnorm(n = 200, mean = c(10, 2), sigma = matrix(c(1, 0.9, 0.9, 2), 2, 2))
contour <- dqcontour(u = 0.9, X = X, weighted = FALSE)

plot(X, cex = 0.5)
points(rbind(contour, contour[1,]), col = 2, lwd = 3, type = "l")

}
